+++ R/cansim.R
@@ -1,35 +1,3 @@
-
-#' Retrieves a Statistics Canada data table using CANSIM code or NDM table number
-#'
-#' Retrieves a data table as a tidy dataframe using either an old-style CANSIM code or a new-format NDM table number. This function will automatically convert old-style CANSIM codes into their new equivalents. Retrieved table data is cached for the duration of the current R session only by default.
-#'
-#' Deprecated, use `get_cansim` instead. This will be removed in future releases.
-#'
-#' @param cansimTableNumber the table number to load, accepts old CANSIM or new NDM table numbers
-#' @param language \code{"en"} or \code{"english"} for English and \code{"fr"} or \code{"french"} for French language versions (default is set to English)
-#' @param refresh (Optional) When set to \code{TRUE}, forces a reload of data table (default is \code{FALSE})
-#' @param timeout (Optional) Timeout in seconds for downloading cansim table to work around scenarios where StatCan servers drop the network connection.
-#  Set to higher values for large tables and slow network connection. (Default is \code{200}).
-#'
-#' @return A tibble with the StatCan table data
-#'
-#' @examples
-#' \dontrun{
-#' # Retrieve a table with an NDM code
-#' get_cansim("34-10-0013")
-#' # Retrieve a table with an old-style CANSIM code
-#' get_cansim("026-0018")
-#' }
-#'
-#' @keywords internal
-#' @export
-get_cansim_ndm <- function(cansimTableNumber, language="english", refresh=FALSE, timeout = 200){
-  warning("The `get_cansim_ndm` method is deprecated and will be removed in future releases, please use `get_cansim` instead.")
-  get_cansim(cleaned_ndm_table_number(cansimTableNumber), language, refresh, timeout = timeout)
-}
-
-
-
 #' Normalize retrieved data table values to appropriate scales
 #'
 #' Facilitates working with Statistics Canada data table values retrieved using the package by setting all units to counts/dollars instead of millions, etc. If "replacement_value" is not set, it will replace the \code{VALUE} field with normalized values and drop the \code{scale} column. Otherwise it will keep the scale columns and create a new column named replacement_value with the normalized value. It will attempt to parse the \code{REF_DATE} field and create an R date variable. This is currently experimental.
@@ -129,10 +97,10 @@ normalize_cansim_values <- function(data, replacement_value=NA, normalize_percen
                                 paste0("Nom g",intToUtf8(0x00E9),"ographique"),
                                 paste0("G",intToUtf8(0x00E9),"ographie d'origine")))
 
-  if (!is.null(cansimTableNumber)) {
-    base_table <- naked_ndm_table_number(cansimTableNumber)
-    path <- paste0(base_path_for_table_language(cansimTableNumber,language),".zip")
-    data_path <- paste0(base_path_for_table_language(cansimTableNumber,language),".Rda")
+  base_table <- naked_ndm_table_number(cansimTableNumber)
+  path <- paste0(base_path_for_table_language(cansimTableNumber,language),".zip")
+  data_path <- paste0(base_path_for_table_language(cansimTableNumber,language),".Rda")
+  if (!is.null(cansimTableNumber) && file.exists(paste0(data_path,"2"))) {
     meta2 <- readRDS(paste0(data_path,"2"))
     dimension_name_column <- ifelse(cleaned_language=="eng","Dimension name","Nom de la dimension")
     fields <- pull(meta2,  dimension_name_column)
@@ -298,105 +266,7 @@ cansim_old_to_new <- function(oldCansimTableNumber){
 }
 
 
-#' Parse metadata
-#' @param meta the raw metadata table
-#' @param data_path base path to save parsed metadata
-#'
-#' @return NULL
-#' @keywords internal
-parse_metadata <- function(meta,data_path){
-  cleaned_language <- basename(data_path) %>% gsub("^.+-|\\..+$","",.)
-  cube_title_column <- ifelse(cleaned_language=="eng","Cube Title","Titre du cube")
-  dimension_id_column <- ifelse(cleaned_language=="eng","Dimension ID",paste0("Num",intToUtf8(0x00E9),"ro d'identification de la dimension"))
-  dimension_name_column <- ifelse(cleaned_language=="eng","Dimension name","Nom de la dimension")
-  classification_code_column <- ifelse(cleaned_language=="eng","Classification Code","Code sur la classification")
-  member_name_column <- ifelse(cleaned_language=="eng","Member Name","Nom du membre")
-  geography_column <- ifelse(cleaned_language=="eng","Geography",paste0("G",intToUtf8(0x00E9),"ographie"))
-  data_geography_column <- ifelse(cleaned_language=="eng","GEO",paste0("G",intToUtf8(0x00C9),"O"))
-  symbol_legend_grepl_field <- ifelse(cleaned_language=="eng","Symbol Legend",paste0("L",intToUtf8(0x00E9),"gende Symbole"))
-  survey_code_grepl_field <- ifelse(cleaned_language=="eng","Survey Code",paste0("Code d'enqu",intToUtf8(0x00EA),"te"))
-  subject_code_grepl_field <- ifelse(cleaned_language=="eng","Subject Code","Code du sujet")
-  note_id_grepl_field <- ifelse(cleaned_language=="eng","Note ID",paste0("Num",intToUtf8(0x00E9),"ro d'identification de la note"))
-  correction_id_grepl_field <- ifelse(cleaned_language=="eng","Correction ID",paste0("Num",intToUtf8(0x00E9),"ro d'identification de la correction"))
-  member_id_column <- ifelse(cleaned_language=="eng","Member ID",paste0("Num",intToUtf8(0x00E9),"ro d'identification du membre"))
-  parent_member_id_column <- ifelse(cleaned_language=="eng","Parent Member ID",paste0("Num",intToUtf8(0x00E9),"ro d'identification du membre parent"))
-  hierarchy_column <- ifelse(cleaned_language=="eng","Hierarchy",paste0("Hi",intToUtf8(0x00E9),"rarchie"))
-  exceeded_hierarchy_warning_message <- ifelse(cleaned_language=="eng","Exceeded max depth for hierarchy, hierarchy information may be faulty.",
-                                               paste0("Profondeur maximale d",intToUtf8(0x00E9),"pass",intToUtf8(0x00E9),"e pour la hi",intToUtf8(0x00E9),"rarchie, les informations de hi",intToUtf8(0x00E9),"rarchie peuvent ",intToUtf8(0x00EA),"tre erron",intToUtf8(0x00E9),"es."))
-  classification_code_prefix <- ifelse(cleaned_language=="eng","Classification Code for","Code de classification pour")
-  hierarchy_prefix <- ifelse(cleaned_language=="eng","Hierarchy for",paste0("Hi",intToUtf8(0x00E9),"rarchie pour"))
 
-  m<-suppressWarnings(setdiff(grep(dimension_id_column,meta[[cube_title_column]]),nrow(meta)))
-  m<-NULL
-
-  cut_indices <- setdiff(grep(dimension_id_column,meta[[cube_title_column]]),nrow(meta))
-  cut_indices <- c(cut_indices,grep(symbol_legend_grepl_field,meta[[cube_title_column]])) %>% sort()
-  meta1 <- meta[seq(1,cut_indices[1]-1),]
-  saveRDS(meta1,file=paste0(data_path,"1"))
-  names2 <- meta[cut_indices[1],]  %>%
-    dplyr::select_if(function(d)sum(!is.na(d)) > 0) %>%
-    as.character()
-  meta2 <- meta[seq(cut_indices[1]+1,cut_indices[2]-1),seq(1,length(names2))] %>%
-    rlang::set_names(names2)
-  saveRDS(meta2,file=paste0(data_path,"2"))
-  names3 <- meta[cut_indices[2],]  %>%
-    dplyr::select_if(function(d)sum(!is.na(d)) > 0) %>%
-    as.character()
-  meta3 <- meta[seq(cut_indices[2]+1,cut_indices[3]-1),seq(1,length(names3))] %>%
-    rlang::set_names(names3)
-  correction_index <- grep(correction_id_grepl_field,meta[[cube_title_column]])
-  if (length(correction_index)==0) correction_index=nrow(meta)
-  additional_indices=c(grep(survey_code_grepl_field,meta[[cube_title_column]]),
-                       grep(subject_code_grepl_field,meta[[cube_title_column]]),
-                       grep(note_id_grepl_field,meta[[cube_title_column]]),
-                       correction_index)
-  saveRDS(meta[seq(additional_indices[1]+1,additional_indices[2]-1),c(1,2)] %>%
-            rlang::set_names(meta[additional_indices[1],c(1,2)]) ,file=paste0(data_path,"3"))
-  saveRDS(meta[seq(additional_indices[2]+1,additional_indices[3]-1),c(1,2)] %>%
-            rlang::set_names(meta[additional_indices[2],c(1,2)]) ,file=paste0(data_path,"4"))
-  if (length(additional_indices)>3)
-    saveRDS(meta[seq(additional_indices[3]+1,additional_indices[4]-1),c(1,2)] %>%
-              rlang::set_names(meta[additional_indices[3],c(1,2)]) ,file=paste0(data_path,"5"))
-
-  add_hierarchy <- function(meta_x){
-    parent_lookup <- rlang::set_names(meta_x[[parent_member_id_column]],meta_x[[member_id_column]])
-    current_top <- function(c){
-      strsplit(c,"\\.") %>%
-        purrr::map(dplyr::first) %>%
-        unlist
-    }
-    parent_for_current_top <- function(c){
-      as.character(parent_lookup[current_top(c)])
-    }
-    meta_x <- meta_x %>%
-      dplyr::mutate(!!as.name(hierarchy_column):=.data[[member_id_column]])
-    added=TRUE
-    max_depth=100
-    count=0
-    while (added & count<max_depth) { # generate hierarchy data from member id and parent member id data
-      old <- meta_x[[hierarchy_column]]
-      meta_x <- meta_x %>%
-        dplyr::mutate(p=parent_for_current_top(.data[[hierarchy_column]])) %>%
-        dplyr::mutate(!!as.name(hierarchy_column):=ifelse(is.na(.data$p),.data[[hierarchy_column]],paste0(.data$p,".",.data[[hierarchy_column]]))) %>%
-        dplyr::select(-.data$p)
-      added <- sum(old != meta_x[[hierarchy_column]])>0
-      count=count+1
-    }
-    if (added) warning(exceeded_hierarchy_warning_message)
-    meta_x
-  }
-  column_names <- dplyr::pull(meta2,dimension_name_column)
-  for (column_index in seq(1:nrow(meta2))) { # iterate through columns for which we have meta data
-    column <- meta2[column_index,]
-    is_geo_column <- grepl(geography_column,column[[dimension_name_column]]) &  !(column[[dimension_name_column]] %in% column_names)
-    meta_x <- meta3 %>%
-      dplyr::filter(.data[[dimension_id_column]]==column[[dimension_id_column]]) %>%
-      add_hierarchy() %>%
-      mutate(name=ifelse(is.na(!!as.name(classification_code_column)) | is_geo_column,!!as.name(member_name_column),paste0(!!as.name(member_name_column)," ",!!as.name(classification_code_column))))
-    saveRDS(meta_x,file=paste0(data_path,"_column_",column[[dimension_name_column]]))
-  }
-  NULL
-}
 
 #' Fold in metadata and for selected columns
 #' @param data A tibble with StatCan table data as e.g. returned by \code{get_cansim}.
@@ -452,7 +322,7 @@ fold_in_metadata_for_columns <- function(data,data_path,column_names){
 
     column <- meta2 %>% filter(!!as.name(dimension_id_column)==column_index)
     is_geo_column <- grepl(geography_column,column[[dimension_name_column]]) &  !(column[[dimension_name_column]] %in% names(data))
-    meta_x=readRDS(paste0(data_path,"_column_",column[[dimension_name_column]]))
+    meta_x=readRDS(paste0(data_path,"_column_",column_index))
 
     if (is_geo_column) {
       hierarchy_name <- paste0(hierarchy_prefix," ", data_geography_column)
@@ -654,10 +524,40 @@ get_cansim <- function(cansimTableNumber, language="english", refresh=FALSE, tim
 get_cansim_table_info <- function(cansimTableNumber, language="english", refresh=FALSE, timeout=200){
   cleaned_number <- cleaned_ndm_table_number(cansimTableNumber)
   data_path <- paste0(base_path_for_table_language(cleaned_number,language),".Rda1")
-  if (refresh | !file.exists(data_path)){
-    get_cansim(cleaned_number,language=language,refresh = refresh,timeout=timeout)
+  if (!refresh && file.exists(data_path)){
+    message(paste0("Reading CANSIM NDM product ",cleaned_number," information from cache."))
+    result <- readRDS(file=data_path)
+  } else {
+    # get_cansim(cleaned_number,language=language,refresh = refresh,timeout=timeout)
+    cleaned_language <- cleaned_ndm_language(language)
+    cube_title_column <- ifelse(cleaned_language=="eng","Cube Title","Titre du cube")
+    start_period_column <- ifelse(cleaned_language=="eng","Start Reference Period",paste0("D",intToUtf8(0x00E9),"but de la p",intToUtf8(0x00E9),"riode de r",intToUtf8(0x00E9),"f",intToUtf8(0x00E9),"rence"))
+    end_period_column <- ifelse(cleaned_language=="eng","End Reference Period",paste0("Fin de la p",intToUtf8(0x00E9),"riode de r",intToUtf8(0x00E9),"f",intToUtf8(0x00E9),"rence"))
+    frequency_column <- ifelse(cleaned_language=="eng","Frequency",paste0("Fr",intToUtf8(0x00E9),"quence"))
+    dimension_name_column <- ifelse(cleaned_language=="eng","Dimension name","Nom de la dimension")
+    member_name_column <- ifelse(cleaned_language=="eng","Member Name","Nom du membre")
+    archived_column <- "Archive Status"
+
+    d <- get_cansim_cube_metadata(cansimTableNumber, type="overview",refresh=refresh)
+
+    if (cleaned_language=="fra") {
+      result <- d %>%
+        select(!!cube_title_column:=.data$cubeTitleFr,
+               `Product Id`=.data$productId, `CANSIM Id`=.data$cansimId,
+               !!archived_column:=.data$archiveStatusFr,
+               !!frequency_column:=.data$frequencyCode,
+               !!start_period_column:=.data$cubeStartDate,!!end_period_column:=.data$cubeEndDate)
+    } else {
+      result <- d %>%
+        select(!!cube_title_column:=.data$cubeTitleEn,
+               `Product Id`=.data$productId, `CANSIM Id`=.data$cansimId,
+               !!archived_column:=.data$archiveStatusEn,
+               !!frequency_column:=.data$frequencyCode,
+               !!start_period_column:=.data$cubeStartDate,!!end_period_column:=.data$cubeEndDate)
+
+    }
   }
-  readRDS(file=data_path)
+  result
 }
 
 
@@ -681,10 +581,14 @@ get_cansim_table_info <- function(cansimTableNumber, language="english", refresh
 get_cansim_table_survey <- function(cansimTableNumber, language="english", refresh=FALSE, timeout=200){
   cleaned_number <- cleaned_ndm_table_number(cansimTableNumber)
   data_path <- paste0(base_path_for_table_language(cleaned_number,language),".Rda3")
-  if (refresh | !file.exists(data_path)){
-    get_cansim(cleaned_number,language=language,refresh = refresh, timeout = timeout)
+  if (!refresh && file.exists(data_path)) {
+    result <- readRDS(file=data_path)
+  } else {
+    cleaned_language <- cleaned_ndm_language(language)
+    survey_code_grepl_field <- ifelse(cleaned_language=="eng","Survey Code",paste0("Code d'enqu",intToUtf8(0x00EA),"te"))
+    result<-get_cansim_cube_metadata(cansimTableNumber,type="overview",refresh=refresh) %>% select(!!survey_code_grepl_field:=.data$surveyCode)
   }
-  readRDS(file=data_path)
+  result
 }
 
 #' Retrieve Statistics Canada data table subject detail
@@ -707,10 +611,18 @@ get_cansim_table_survey <- function(cansimTableNumber, language="english", refre
 get_cansim_table_subject <- function(cansimTableNumber, language="english", refresh=FALSE, timeout = 200){
   cleaned_number <- cleaned_ndm_table_number(cansimTableNumber)
   data_path <- paste0(base_path_for_table_language(cleaned_number,language),".Rda4")
-  if (refresh | !file.exists(data_path)){
-    get_cansim(cleaned_number,language=language,refresh = refresh, timeout = timeout)
+  if (!refresh && file.exists(data_path)) {
+    result <- readRDS(file=data_path)
+  } else {
+    cleaned_language <- cleaned_ndm_language(language)
+    subject_code_grepl_field <- ifelse(cleaned_language=="eng","Subject Code","Code du sujet")
+    result<-get_cansim_cube_metadata(cansimTableNumber,type="overview",refresh=refresh) %>%
+      select(.data$subjectCode) %>%
+      mutate(subjectCode=strsplit(.data$subjectCode,", ")) %>%
+      tidyr::unnest_longer(.data$subjectCode) %>%
+      select(!!subject_code_grepl_field:=.data$subjectCode)
   }
-  readRDS(file=data_path)
+  result
 }
 
 #' Retrieve Statistics Canada data table short notes
@@ -733,10 +645,30 @@ get_cansim_table_subject <- function(cansimTableNumber, language="english", refr
 get_cansim_table_short_notes <- function(cansimTableNumber, language="english", refresh=FALSE, timeout = 200){
   cleaned_number <- cleaned_ndm_table_number(cansimTableNumber)
   data_path <- paste0(base_path_for_table_language(cleaned_number,language),".Rda5")
-  if (refresh | !file.exists(data_path)){
-    get_cansim(cleaned_number,language=language,refresh = refresh, timeout = timeout)
+  if (!refresh && file.exists(data_path)) {
+    notes <- readRDS(file=data_path)
+  } else if (!file.exists(data_path)) {
+    notes <- get_cansim_cube_metadata(cansimTableNumber,refresh=refresh,type="notes")
   }
-  readRDS(file=data_path)
+  if (refresh || !file.exists(data_path)){
+    notes <- get_cansim_cube_metadata(cansimTableNumber,refresh=refresh,type="notes")
+    cleaned_language <- cleaned_ndm_language(language)
+    note_id_grepl_field <- ifelse(cleaned_language=="eng","Note ID",paste0("Num",intToUtf8(0x00E9),"ro d'identification de la note"))
+
+    if (cleaned_language=="fra") {
+      notes <- notes %>%
+        select(.data$footnoteId,Note=.data$footnotesFr)
+    } else {
+      notes <- notes %>%
+        select(.data$footnoteId,Note=.data$footnotesEn)
+    }
+    notes <- notes %>%
+      unique() %>%
+      rename(!!note_id_grepl_field:=.data$footnoteId)
+    #get_cansim(cleaned_number,language=language,refresh = refresh, timeout = timeout)
+  }
+
+  notes
 }
 
 #' Retrieve Statistics Canada data table column list
@@ -759,10 +691,28 @@ get_cansim_table_short_notes <- function(cansimTableNumber, language="english", 
 get_cansim_column_list <- function(cansimTableNumber, language="english", refresh=FALSE, timeout= 200){
   cleaned_number <- cleaned_ndm_table_number(cansimTableNumber)
   data_path <- paste0(base_path_for_table_language(cleaned_number,language),".Rda2")
-  if (refresh | !file.exists(data_path)){
-    get_cansim(cleaned_number,language=language,refresh = refresh, timeout = timeout)
+  if (!refresh && file.exists(data_path)) {
+    result <- readRDS(file=data_path)
+  } else {
+    cleaned_language <- cleaned_ndm_language(language)
+    dimension_id_column <- ifelse(cleaned_language=="eng","Dimension ID",paste0("Num",intToUtf8(0x00E9),"ro d'identification de la dimension"))
+    dimension_name_column <- ifelse(cleaned_language=="eng","Dimension name","Nom de la dimension")
+    d <- get_cansim_cube_metadata(cansimTableNumber,type="members",refresh=refresh)
+
+    if (cleaned_language=="fra") {
+     result <- d %>%
+      select(!!dimension_id_column:=.data$dimensionPositionId,!!dimension_name_column:=.data$dimensionNameFr) %>%
+       unique()
+    } else {
+      result <- d %>%
+        select(!!dimension_id_column:=.data$dimensionPositionId,!!dimension_name_column:=.data$dimensionNameEn)  %>%
+        unique()
+    }
+    #get_cansim(cleaned_number,language=language,refresh = refresh, timeout = timeout)
+    #result <- readRDS(file=data_path)
   }
-  readRDS(file=data_path)
+
+  result
 }
 
 #' Retrieve Statistics Canada data table categories for a specific column
@@ -785,15 +735,57 @@ get_cansim_column_list <- function(cansimTableNumber, language="english", refres
 #' @export
 get_cansim_column_categories <- function(cansimTableNumber, column, language="english", refresh=FALSE, timeout = 200){
   cleaned_number <- cleaned_ndm_table_number(cansimTableNumber)
+  cleaned_language <- cleaned_ndm_language(language)
   data_path <- paste0(base_path_for_table_language(cleaned_number,language),".Rda2")
-  if (refresh | !file.exists(data_path)){
-    get_cansim(cleaned_number,language=language,refresh = refresh, timeout = 200)
+  if (!refresh && file.exists(data_path)) {
+    meta2 <- readRDS(data_path)
+    dimension_name_column <- ifelse(cleaned_language=="eng","Dimension name","Nom de la dimension")
+    dimension_id_column <- ifelse(cleaned_language=="eng","Dimension ID",paste0("Num",intToUtf8(0x00E9),"ro d'identification de la dimension"))
+    column_index <- meta2 %>%
+      dplyr::filter(!!as.name(dimension_name_column) == column) %>%
+      dplyr::pull(!!as.name(dimension_id_column))
+    data_path <- paste0(base_path_for_table_language(cleaned_number,language),".Rda_column_",column_index)
+    if (!file.exists(data_path)){
+      stop(paste0("Unkown column ",column))
+    }
+    result <- readRDS(file=data_path)
+  } else {
+    cleaned_language <- cleaned_ndm_language(language)
+    dimension_id_column <- ifelse(cleaned_language=="eng","Dimension ID",paste0("Num",intToUtf8(0x00E9),"ro d'identification de la dimension"))
+    dimension_name_column <- ifelse(cleaned_language=="eng","Dimension name","Nom de la dimension")
+    member_name_column <- ifelse(cleaned_language=="eng","Member Name","Nom du membre")
+    member_id_column <- ifelse(cleaned_language=="eng","Member ID",paste0("Num",intToUtf8(0x00E9),"ro d'identification du membre"))
+    parent_member_id_column <- ifelse(cleaned_language=="eng","Parent Member ID",paste0("Num",intToUtf8(0x00E9),"ro d'identification du membre parent"))
+    hierarchy_column <- ifelse(cleaned_language=="eng","Hierarchy",paste0("Hi",intToUtf8(0x00E9),"rarchie"))
+    terminated_column <- ifelse(cleaned_language=="eng","Terminated",paste0("Termin",intToUtf8(0x00E9)))
+    exceeded_hierarchy_warning_message <- ifelse(cleaned_language=="eng","Exceeded max depth for hierarchy, hierarchy information may be faulty.",
+                                                 paste0("Profondeur maximale d",intToUtf8(0x00E9),"pass",intToUtf8(0x00E9),"e pour la hi",intToUtf8(0x00E9),"rarchie, les informations de hi",intToUtf8(0x00E9),"rarchie peuvent ",intToUtf8(0x00EA),"tre erron",intToUtf8(0x00E9),"es."))
+    d <- get_cansim_cube_metadata(cansimTableNumber,type="members",refresh=refresh)
+
+    if (cleaned_language=="fra") {
+      result <- d %>%
+        select(!!dimension_id_column:=.data$dimensionPositionId,!!dimension_name_column:=.data$dimensionNameFr,
+               !!member_id_column:=.data$memberId, !!member_name_column:=.data$memberNameFr,
+               !!parent_member_id_column:=.data$parentMemberId, !!terminated_column:=.data$terminated)
+    } else {
+      result <- d %>%
+        select(!!dimension_id_column:=.data$dimensionPositionId,!!dimension_name_column:=.data$dimensionNameEn,
+               !!member_id_column:=.data$memberId, !!member_name_column:=.data$memberNameEn,
+               !!parent_member_id_column:=.data$parentMemberId, !!terminated_column:=.data$terminated)
+    }
+    result <- result %>%
+      filter(!!as.name(dimension_name_column)==column) %>%
+      add_hierarchy(parent_member_id_column=parent_member_id_column,
+                    member_id_column=member_id_column,
+                    hierarchy_column=hierarchy_column,
+                    exceeded_hierarchy_warning_message=exceeded_hierarchy_warning_message)
+
+    if (nrow(result)==0){
+      stop(paste0("Unkown column ",column))
+    }
   }
-  data_path <- paste0(base_path_for_table_language(cleaned_number,language),".Rda_column_",column)
-  if (!file.exists(data_path)){
-    stop(paste0("Unkown column ",column))
-  }
-  readRDS(file=data_path)
+
+  result
 }
 
 #' Retrieve Statistics Canada data table overview text
@@ -813,7 +805,7 @@ get_cansim_column_categories <- function(cansimTableNumber, column, language="en
 #' @export
 get_cansim_table_overview <- function(cansimTableNumber, language="english", refresh=FALSE){
   cansimTableNumber <- cleaned_ndm_table_number(cansimTableNumber)
-  info <- cansim::get_cansim_table_info(cansimTableNumber,language=language,refresh=refresh)
+  info <- get_cansim_table_info(cansimTableNumber,language=language,refresh=refresh)
   #refresh=FALSE
   cleaned_language <- cleaned_ndm_language(language)
   cube_title_column <- ifelse(cleaned_language=="eng","Cube Title","Titre du cube")
@@ -964,63 +956,7 @@ view_cansim_webpage <- function(cansimTableNumber = NULL){
   utils::browseURL(url,browser)
 }
 
-#' Retrieve table metadata from Statistics Canada API
-#'
-#' Retrieves table metadata given an input table number or vector of table numbers using either the new or old table number format. Patience is suggested as the Statistics Canada API can be very slow. The `list_cansim_tables()` function can be used as an alternative to retrieve a (cached) list of CANSIM tables with (more limited) metadata.
-#'
-#' @param cansimTableNumber A new or old CANSIM/NDM table number or a vector of table numbers
-#'
-#' @return a tibble containing the table metadata
-#'
-#' @examples
-#' \dontrun{
-#' get_cansim_cube_metadata("34-10-0013")
-#' }
-#' @export
-get_cansim_cube_metadata <- function(cansimTableNumber){
-  cansimTableNumber <- cleaned_ndm_table_number(cansimTableNumber)
-  table_id <- naked_ndm_table_number(cansimTableNumber)
-  url <- "https://www150.statcan.gc.ca/t1/wds/rest/getCubeMetadata"
-  response <- httr::POST(url,
-                     #body=jsonlite::toJSON(list("productId"=table_id),auto_unbox =TRUE),
-                     body=paste0("[",paste(paste0('{"productId":',table_id,'}'),collapse = ", "),"]"),
-                     encode="json",
-                     httr::add_headers("Content-Type"="application/json")
-  )
-  if (response$status_code!=200) {
-    stop("Problem downloading data, status code ",response$status_code,"\n",httr::content(response))
-  }
-  data <- httr::content(response)
-  data1 <- Filter(function(x)x$status=="SUCCESS",data)
-  data2 <- Filter(function(x)x$status!="SUCCESS",data)
-  if (length(data2)>0) {
-    message(paste0("Failed to load metadata for ",length(data2)," tables "))
-    data2 %>% purrr::map(function(x){
-      message(x$object)
-    })
-  }
-  fields <- c("productId", "cansimId", "cubeTitleEn", "cubeTitleFr", "cubeStartDate", "cubeEndDate", "nbSeriesCube",
-              "nbDatapointsCube",  "archiveStatusCode", "archiveStatusEn",   "archiveStatusFr",   "subjectCode",
-              "surveyCode",  "dimension","releaseTime")
-  # fields <- c("productId", "cansimId", "cubeTitleEn", "cubeTitleFr", "cubeStartDate", "cubeEndDate",
-  #             "frequencyCode", "nbSeriesCube", "nbDatapointsCube",
-  #             "releaseTime", "archiveStatusCode", "archiveStatusEn", "archiveStatusFr",   "subjectCode",
-  #             "surveyCode",  "dimension", "footnote","correctionFootnote", "correction")
-  l <- lapply(fields, function(field){
-    purrr::map(data1,function(d){
-      dd<-d$object[[field]]
-      if (typeof(dd)=="list") dd <- dd %>% unlist %>% as.character() %>% paste(collapse = ",")
-      dd
-      }) %>% as.character()
-  }) %>%
-    purrr::set_names(fields) %>%
-    tibble::as_tibble() %>%
-    dplyr::mutate(productId=cleaned_ndm_table_number(.data$productId)) %>%
-    dplyr::mutate(releaseTime=readr::parse_datetime(.data$releaseTime,
-                                                    format=STATCAN_TIME_FORMAT,
-                                                    locale=readr::locale(tz=STATCAN_TIMEZONE)))
-  l
-}
+
 
 #' Retrieve a Statistics Canada data table URL given a table number
 #'
@@ -1126,19 +1062,47 @@ get_cansim_table_notes <- function(cansimTableNumber,language="en",refresh=FALSE
   note_id_column <- ifelse(cleaned_language=="eng","Note ID",paste0("Num",intToUtf8(0x00E9),"ro d'identification de la note"))
   notes <- get_cansim_table_short_notes(cansimTableNumber,language=language,refresh=refresh,timeout=timeout)
   columns <- get_cansim_column_list(cansimTableNumber,language=language)
-  full_notes <- columns %>%
-    select(!!as.name(dimension_name_column),!!note_id_column:=!!as.name(dimenion_note_column)) %>%
-    bind_rows(
-      pull(.,dimension_name_column) %>% lapply(function(c) {
-        get_cansim_column_categories(cansimTableNumber,column=c,language = language) %>%
-          mutate(!!dimension_name_column:=c) %>%
-          select(!!as.name(dimension_name_column),!!as.name(member_name_column),
-                 !!note_id_column:=!!as.name(member_note_column))
-      }) %>%
-        bind_rows) %>%
-    filter(!is.na(!!as.name(note_id_column))) %>%
-    full_join(notes,by=note_id_column) %>%
-    arrange(!!as.name(note_id_column))
+
+  if (dimenion_note_column %in% names(columns)) {
+    full_notes <- columns %>%
+      select(!!as.name(dimension_name_column),!!note_id_column:=!!as.name(dimenion_note_column)) %>%
+      bind_rows(
+        pull(.,dimension_name_column) %>% lapply(function(c) {
+          get_cansim_column_categories(cansimTableNumber,column=c,language = language) %>%
+            mutate(!!dimension_name_column:=c) %>%
+            select(!!as.name(dimension_name_column),!!as.name(member_name_column),
+                   !!note_id_column:=!!as.name(member_note_column))
+        }) %>%
+          bind_rows) %>%
+      filter(!is.na(!!as.name(note_id_column))) %>%
+      mutate(!!note_id_column:=strsplit(!!as.name(note_id_column),";")) %>%
+      tidyr::unnest_longer(!!note_id_column) %>%
+      full_join(notes,by=note_id_column) %>%
+      arrange(!!as.name(note_id_column))
+  } else {
+    full_notes <- get_cansim_cube_metadata(cansimTableNumber,type="notes",refresh=refresh)
+    members <- get_cansim_cube_metadata(cansimTableNumber,type="members",refresh = refresh)
+
+    if (cleaned_language=="fra") {
+      members <- members %>%
+        select(.data$dimensionPositionId,!!dimension_name_column:=.data$dimensionNameFr,
+               .data$memberId,!!member_name_column:=.data$memberNameFr) %>%
+        unique()
+      full_notes <- full_notes %>%
+        select(!!note_id_column:=.data$footnoteId,Note=.data$footnotesFr,.data$dimensionPositionId, .data$memberId) %>%
+        left_join(members,by=c("dimensionPositionId","memberId")) %>%
+        select(-.data$dimensionPositionId,-.data$memberId)
+    } else {
+      members <- members %>%
+        select(.data$dimensionPositionId,!!dimension_name_column:=.data$dimensionNameEn,
+               .data$memberId,!!member_name_column:=.data$memberNameEn) %>%
+        unique()
+      full_notes <- full_notes %>%
+        select(!!note_id_column:=.data$footnoteId,Note=.data$footnotesFr,.data$dimensionPositionId, .data$memberId) %>%
+        left_join(members,by=c("dimensionPositionId","memberId")) %>%
+        select(-.data$dimensionPositionId,-.data$memberId)
+    }
+  }
   full_notes
 }