+++ NEWS.md
@@ -1,8 +1,9 @@
 # cansim 0.3.17
 ## Minor changes
-* fix problem with reading french tables released by the census division
+* fix problem with reading French tables released by the census division
 * restore original column order after converting to factors
 * convery geography column to factor if available
+* fix problem with `add_provincial_abbreviations` that could lead to mislabelling of provinces in some cases
 
 # cansim 0.3.16
 ## Minor changes